Fixed another false positive in Interval.intersects.
If i1 = (a, b) and i2 = (b, c), then the two intervals don't
intersect (i1 doesn't contain b because end is exclusive).
there is a test that i1.intersects(i2) is false, and that
test is green. However, due to an off-by-one error,
i2.intersects(i1) was true before this commit.