Added a hamcrest matcher for aggregations.
This makes it easy to create a bunch of simple tests for the
aggregations. You can just use it like this:

> assertThat(foo, aggregatesValues(1, 3, 4).into(9));

On failure, the matcher will output the input array, the expected
value and the value it got. The matcher won't call reset in any
way, because that would make testing reset itself harder.